requirements.txt
torch
torchmetrics
pandas
numpy
transformers
setuptools
python-dotenv
scikit-learn
mlflow
deep_translator
langdetect
streamlit
pymongo
tensorflow